web_application/requirements/development.txt
-r production.txt
model-mommy
isort
mypy
behave
behave-django
pytest
pytest-django
pylint
flake8
black
coverage
safety
bandit
